Reimplemented plotting + other modifications. (#60)
- histograms.cfg: added blind_ranges for mva_score.
- plot_setup.cfg: defined default setup for stacked plot with ratio.
- sources.cfg:
  - using shortened category names;
  - fixed mistype in mva_hmANkin_noSVfit vars.
- AnaTuple.h, MvaReader.h: fixed double -> float warnings.
- AnalysisCategories.h: using short WP names in EventCategory.
- EventAnalyzerCore: disabled copy constructor.
- SampleDescriptor.h: added plot_cfg and plot_page_opt parameters.
- StackedPlotsProducer: using new plotting code.
- CMakeLists.txt: disabled precompiled headers.
- SMFileMerger.cxx: fixed bug in weight calculation.
- bbmumuAnalyzer.cxx, HHReweight5D.h, DY_estimation.cxx, MvaEvaluation.cxx, MvaTraining.cxx: fixed compilation warnings.

+++ Analysis/include/MvaReader.h
@@ -94,16 +94,16 @@ public:
 
         const auto& met = eventbase.GetMET();
 
-        dphi_mumet = std::abs(ROOT::Math::VectorUtil::DeltaPhi(t1.GetMomentum(), met.GetMomentum()));
-        dphi_metsv = std::abs(ROOT::Math::VectorUtil::DeltaPhi(Htt_sv, met.GetMomentum()));
-        dR_bb = std::abs(ROOT::Math::VectorUtil::DeltaR(b1.GetMomentum(), b2.GetMomentum()));
-        dR_bbbb = ROOT::Math::VectorUtil::DeltaR(b1.GetMomentum(), b2.GetMomentum())*Hbb.GetMomentum().Pt();
-        dR_taumu = std::abs(ROOT::Math::VectorUtil::DeltaR(t1.GetMomentum(), t2.GetMomentum()));
-        dR_taumusvfit = ROOT::Math::VectorUtil::DeltaR(t1.GetMomentum(), t2.GetMomentum())*Htt.Pt();
+        dphi_mumet = static_cast<float>(std::abs(ROOT::Math::VectorUtil::DeltaPhi(t1.GetMomentum(), met.GetMomentum())));
+        dphi_metsv = static_cast<float>(std::abs(ROOT::Math::VectorUtil::DeltaPhi(Htt_sv, met.GetMomentum())));
+        dR_bb = static_cast<float>(std::abs(ROOT::Math::VectorUtil::DeltaR(b1.GetMomentum(), b2.GetMomentum())));
+        dR_bbbb = static_cast<float>(ROOT::Math::VectorUtil::DeltaR(b1.GetMomentum(), b2.GetMomentum())*Hbb.GetMomentum().Pt());
+        dR_taumu = static_cast<float>(std::abs(ROOT::Math::VectorUtil::DeltaR(t1.GetMomentum(), t2.GetMomentum())));
+        dR_taumusvfit = static_cast<float>(ROOT::Math::VectorUtil::DeltaR(t1.GetMomentum(), t2.GetMomentum())*Htt.Pt());
         mT1 = static_cast<float>(Calculate_MT(t1.GetMomentum(), met.GetMomentum()));
         mT2 = static_cast<float>(Calculate_MT(t2.GetMomentum(), met.GetMomentum()));
-        dphi_bbmet = std::abs(ROOT::Math::VectorUtil::DeltaPhi(Hbb.GetMomentum(), met.GetMomentum()));
-        dphi_bbsv = std::abs(ROOT::Math::VectorUtil::DeltaPhi(Hbb.GetMomentum(), Htt_sv));
+        dphi_bbmet = static_cast<float>(std::abs(ROOT::Math::VectorUtil::DeltaPhi(Hbb.GetMomentum(), met.GetMomentum())));
+        dphi_bbsv = static_cast<float>(std::abs(ROOT::Math::VectorUtil::DeltaPhi(Hbb.GetMomentum(), Htt_sv)));
     }
     virtual double Evaluate() override { return reader->EvaluateMVA(method_name); }
